move type resolver
