+++ src/Shared/Reflection/TypeResolver/AsIsResolver.php
@@ -2,10 +2,9 @@
 
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ArgumentResolver;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
-final class AsIsResolver implements ArgumentResolver
+final class AsIsResolver implements TypeResolver
 {
     public function canResolve(Type $type, mixed $value): bool
     {
