+++ src/Shared/Reflection/TypeResolver/MapResolver.php
@@ -2,11 +2,10 @@
 
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ArgumentResolver;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\ArrayMap;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
-final class MapResolver implements ArgumentResolver
+final class MapResolver implements TypeResolver
 {
     public const SUPPORT_TYPE = ArrayMap::class;
